backend/requirements.txt
Django
psycopg2-binary
graphene-django
django-cors-headers
djangorestframework
pydub
audioop-lts
numpy
